+++ 0x1A-hash_tables/5-hash_table_print.c
@@ -7,7 +7,7 @@
  **/
 void hash_table_print(const hash_table_t *ht)
 {
-	unsigned int i = 0, enc = 0;
+	unsigned long int i = 0, enc = 0;
 	hash_node_t *ptr;
 
 	if (!ht)
